fix macOS CI: silence non-OpenMP warnings and pick venv numpy
clang on macOS builds without OpenMP, exposing four kinds of warnings
that gcc/clang-on-Linux happen to elide:

  - ts_ns() in qblas_cpu.c is only referenced from a _OPENMP-gated
    overhead probe; guard it so it doesn't compile as dead code.
  - level1 threshold comparisons compared a signed int against a size_t
    field; widen to size_t.
  - cblas_qger / trsm_left_diag / trmm_left_diag declared nthreads
    outside their _OPENMP block; move the declaration inside.

test_against_numpy was failing on macOS because find_package(Python3)
captured the framework python at configure time and missed the venv
where numpy is installed. The CI now exports VIRTUAL_ENV alongside the
PATH update, and tests/CMakeLists.txt checks 'import numpy' at
configure time, skipping the test (with a status message) if numpy is
unreachable from the chosen interpreter.

+++ src/level1/qblas_level1.c
@@ -15,7 +15,7 @@ Sleef_quad cblas_qdot(int n,
     if (n <= 0) return QBLAS_ZERO;
     if (incx == 1 && incy == 1) {
         int nthreads = qblas_resolve_threads((size_t)n, 1);
-        if (nthreads > 1 && n >= qblas_tune()->l1_thread_threshold) {
+        if (nthreads > 1 && (size_t)n >= qblas_tune()->l1_thread_threshold) {
 #ifdef _OPENMP
             Sleef_quad totals[256] = {0};
             if (nthreads > 256) nthreads = 256;
@@ -52,7 +52,7 @@ Sleef_quad cblas_qasum(int n, const Sleef_quad *x, int incx) {
     if (n <= 0) return QBLAS_ZERO;
     if (incx == 1) {
         int nthreads = qblas_resolve_threads((size_t)n, 1);
-        if (nthreads > 1 && n >= qblas_tune()->l1_thread_threshold) {
+        if (nthreads > 1 && (size_t)n >= qblas_tune()->l1_thread_threshold) {
 #ifdef _OPENMP
             Sleef_quad totals[256] = {0};
             if (nthreads > 256) nthreads = 256;
@@ -89,7 +89,7 @@ void cblas_qaxpy(int n, Sleef_quad alpha,
     if (n <= 0 || qiszero(alpha)) return;
     if (incx == 1 && incy == 1) {
         int nthreads = qblas_resolve_threads((size_t)n, 2);
-        if (nthreads > 1 && n >= qblas_tune()->l1_thread_threshold) {
+        if (nthreads > 1 && (size_t)n >= qblas_tune()->l1_thread_threshold) {
 #ifdef _OPENMP
             #pragma omp parallel num_threads(nthreads)
             {
@@ -116,7 +116,7 @@ void cblas_qscal(int n, Sleef_quad alpha, Sleef_quad *x, int incx) {
     if (n <= 0) return;
     if (incx == 1) {
         int nthreads = qblas_resolve_threads((size_t)n, 1);
-        if (nthreads > 1 && n >= qblas_tune()->l1_thread_threshold) {
+        if (nthreads > 1 && (size_t)n >= qblas_tune()->l1_thread_threshold) {
 #ifdef _OPENMP
             #pragma omp parallel num_threads(nthreads)
             {
